+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v0.47](https://github.com/arez/arez/tree/v0.47) (2018-01-31)
+[Full Changelog](https://github.com/arez/arez/compare/v0.46...v0.47)
 
 #### Fixed
 * **\[component\]** Make the `AbstractRepository.entities()` method public so that it can be used by extensions.